+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/pinned/list/internal/PinnedMessageListAdapter.kt
@@ -1,11 +1,9 @@
 package io.getstream.chat.android.ui.pinned.list.internal
 
-import android.content.Context
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.getstream.sdk.chat.utils.DateFormatter
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.ui.common.extensions.internal.asMention
@@ -17,12 +15,10 @@ import io.getstream.chat.android.ui.pinned.list.PinnedMessageListView.PinnedMess
 import io.getstream.chat.android.ui.pinned.list.internal.PinnedMessageListAdapter.MessagePreviewViewHolder
 
 internal class PinnedMessageListAdapter(
-    context: Context,
     private val chatDomain: ChatDomain,
 ) : ListAdapter<Message, MessagePreviewViewHolder>(MessageDiffCallback) {
 
     private var pinnedMessageSelectedListener: PinnedMessageSelectedListener? = null
-    private var dateFormatter = DateFormatter.from(context)
 
     var messagePreviewStyle: MessagePreviewStyle? = null
 
@@ -31,8 +27,6 @@ internal class PinnedMessageListAdapter(
             .inflate(parent.streamThemeInflater, parent, false)
             .let { binding ->
                 messagePreviewStyle?.let(binding.root::styleView)
-                binding.root.dateFormatter = dateFormatter
-
                 MessagePreviewViewHolder(binding)
             }
     }